Move fvm_utils/primitives to ipc/types + cleanup
