--- fvm-utils/.gitignore
@@ -1,3 +1,0 @@
-/target
-/Cargo.lock
-.idea
