--- fvm-utils/example/Cargo.toml
@@ -1,36 +1,0 @@
-[package]
-edition = "2021"
-name = "fil_actor_example"
-version = "0.1.0"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-[dependencies]
-fil_actors_runtime = {path = "../runtime", features = ["test_utils", "fil-actor"]}
-primitives = {path = "../primitives"}
-
-frc42_dispatch = "3.2.0"
-fvm_ipld_blockstore = "0.1.1"
-fvm_ipld_encoding = "0.3.3"
-fvm_ipld_hamt = "0.5.1"
-fvm_shared = {version = "=3.2.0", default-features = false}
-
-anyhow = "1.0.56"
-cid = {version = "0.8.3", default-features = false, features = ["serde-codec"]}
-log = "0.4.14"
-num-derive = "0.3.3"
-num-traits = "0.2.14"
-serde = {version = "1.0.136", features = ["derive"]}
-
-[dev-dependencies]
-base64 = "0.13.0"
-blake2b_simd = "1.0.0"
-libsecp256k1 = {version = "0.7"}
-num-traits = "0.2"
-rand = "0.8"
-rand_chacha = "0.3"
-wabt = "0.10.0"
-wasmtime = "0.35.2"
-
-[build-dependencies]
-wasm-builder = "3.0.1"
-wasmtime = "0.35.2"
