--- fvm-utils/runtime/Cargo.toml
@@ -1,83 +1,0 @@
-[package]
-authors = ["ChainSafe Systems <info@chainsafe.io>", "Protocol Labs", "Filecoin Core Devs"]
-description = "System actors for the Filecoin protocol"
-edition = "2021"
-license = "MIT OR Apache-2.0"
-name = "fil_actors_runtime"
-repository = "https://github.com/filecoin-project/builtin-actors"
-version = "0.0.1"
-
-[dependencies]
-base64 = { workspace = true }
-byteorder = "1.4.3"
-castaway = "0.2.2"
-cid = { workspace = true, default-features = false, features = ["serde-codec"]}
-frc42_dispatch = "3.0.0"
-fvm_ipld_amt = { workspace = true, features = ["go-interop"]}
-fvm_ipld_hamt = { workspace = true }
-fvm_shared = { workspace = true }
-indexmap = {version = "1.8.0", features = ["serde-1"]}
-integer-encoding = { workspace = true, default-features = false}
-lazy_static = { workspace = true }
-log = { workspace = true }
-num-derive = { workspace = true }
-num-traits = { workspace = true }
-serde = { workspace = true, features = ["derive"]}
-thiserror = { workspace = true }
-unsigned-varint = "0.7.1"
-# enforce wasm compat
-anyhow = { workspace = true }
-blake2b_simd = { workspace = true }
-fvm_ipld_blockstore = { workspace = true }
-fvm_ipld_encoding = { workspace = true }
-fvm_sdk = { workspace = true, optional = true }
-getrandom = {version = "0.2.3", features = ["js"]}
-hex = { workspace = true, optional = true}
-itertools = "0.10"
-multihash = { workspace = true, default-features = false}
-paste = { workspace = true }
-rand = { workspace = true }
-regex = { workspace = true }
-serde_repr = "0.1.8"
-serde_tuple = { workspace = true }
-sha2 = { workspace = true }
-
-[dev-dependencies]
-derive_builder = "0.10.2"
-hex = { workspace = true }
-rand = { workspace = true }
-
-[features]
-default = []
-fil-actor = ["fvm_sdk"]
-
-# Enable 2k sectors
-sector-2k = []
-# Enable 8m sectors
-sector-8m = []
-# Enable 512m sectors
-sector-512m = []
-# Enable 32g sectors
-sector-32g = []
-# Enable 64g sectors
-sector-64g = []
-# Enable the "default" sectors types.
-sector-default = ["sector-64g", "sector-32g"]
-
-# Reduce min verified deal size from 1MiB to 256B
-small-deals = []
-# Reduce pre-commit delay from 150 epochs, to 10.
-short-precommit = []
-
-# Lower the minimum power requirement to 2k
-min-power-2k = []
-# Lower the minimum power requirement to 2g
-min-power-2g = []
-
-# no collateral for deals (for testing)
-no-provider-deal-collateral = []
-
-# fake proofs (for testing)
-fake-proofs = []
-
-test_utils = ["hex", "multihash/sha2"]
