--- fvm-utils/runtime/README.md
@@ -1,3 +1,0 @@
-# Runtime
-This crate contains the `runtime` wrapper for `fvm`. It provides a wrapper around low level 
-`fvm` calls, providing functions such as transactions and verifications.
