--- fvm-utils/runtime/src/builtin/network.rs
@@ -1,15 +1,0 @@
-// Copyright 2019-2022 ChainSafe Systems
-// SPDX-License-Identifier: Apache-2.0, MIT
-
-use fvm_shared::clock::EPOCH_DURATION_SECONDS;
-pub use fvm_shared::BLOCKS_PER_EPOCH as EXPECTED_LEADERS_PER_EPOCH;
-
-pub const SECONDS_IN_HOUR: i64 = 3600;
-pub const SECONDS_IN_DAY: i64 = 86400;
-pub const SECONDS_IN_YEAR: i64 = 31556925;
-pub const EPOCHS_IN_HOUR: i64 = SECONDS_IN_HOUR / EPOCH_DURATION_SECONDS;
-pub const EPOCHS_IN_DAY: i64 = SECONDS_IN_DAY / EPOCH_DURATION_SECONDS;
-pub const EPOCHS_IN_YEAR: i64 = SECONDS_IN_YEAR / EPOCH_DURATION_SECONDS;
-
-// 1 NanoFIL
-pub const ONE_NANO_FIL: u64 = 10u64.pow(9);
