--- fvm-utils/runtime/src/util/mod.rs
@@ -1,15 +1,0 @@
-// Copyright 2019-2022 ChainSafe Systems
-// SPDX-License-Identifier: Apache-2.0, MIT
-
-pub use self::downcast::*;
-pub use self::message_accumulator::MessageAccumulator;
-pub use self::multimap::*;
-pub use self::set::Set;
-pub use self::set_multimap::SetMultimap;
-
-pub mod cbor;
-mod downcast;
-mod message_accumulator;
-mod multimap;
-mod set;
-mod set_multimap;
