+++ ipc/Makefile
@@ -1,6 +1,6 @@
 .PHONY: all build test lint license check-fmt check-clippy diagrams install-infra clean-infra
 
-CRATE   := ipc-wallet ipc-provider ipc-api
+CRATE   := ipc-wallet ipc-provider ipc-api ipc-types
 PACKAGE := $(patsubst %, --package %, $(CRATE))
 
 all: test build
