+++ ipc/api/Cargo.toml
@@ -26,7 +26,7 @@ thiserror = { workspace = true }
 ethers = { workspace = true }
 
 ipc_actors_abis = { workspace = true }
-primitives = { workspace = true }
+ipc-types = { workspace = true }
 
 [dev-dependencies]
 serde_json = { workspace = true }
