+++ ipc/provider/Cargo.toml
@@ -40,7 +40,7 @@ fvm_shared = { workspace = true }
 fil_actors_runtime = { workspace = true }
 fvm_ipld_encoding = { workspace = true }
 
-primitives = { workspace = true }
+ipc-types = { workspace = true }
 ipc-wallet = { workspace = true, features = ["with-ethers"] }
 ipc-api = { workspace = true }
 ipc_actors_abis = { workspace = true }
