+++ ipc/provider/src/config/serialize.rs
@@ -6,7 +6,7 @@ use crate::config::Subnet;
 use anyhow::anyhow;
 use fvm_shared::address::{Address, Payload};
 use ipc_api::subnet_id::SubnetID;
-use primitives::EthAddress;
+use ipc_types::EthAddress;
 use serde::ser::{Error, SerializeSeq};
 use serde::Serializer;
 use std::collections::HashMap;
@@ -68,7 +68,7 @@ mod tests {
     use crate::config::{Config, Subnet};
     use fvm_shared::address::Address;
     use ipc_api::subnet_id::SubnetID;
-    use primitives::EthAddress;
+    use ipc_types::EthAddress;
     use std::str::FromStr;
 
     const STR: &str = r#"
