+++ ipc/types/Cargo.toml
@@ -1,11 +1,11 @@
 [package]
 description = "The primitive types for fvm"
 edition = "2021"
-name = "primitives"
+name = "ipc-types"
 version = "0.1.0"
 
 [dependencies]
-fil_actors_runtime = {path = "../runtime", features = ["fil-actor"]}
+fvm_ipld_amt = { workspace = true, features = ["go-interop"]}
 fvm_ipld_blockstore = { workspace = true }
 fvm_ipld_encoding = { workspace = true }
 fvm_ipld_hamt = { workspace = true }
@@ -22,3 +22,4 @@ num-derive =  { workspace = true }
 num-traits =  { workspace = true }
 serde =  { workspace = true, features = ["derive"]}
 uint = {version = "0.9.3", default-features = false}
+thiserror = { workspace = true }
